API/requirements.txt
anyio==3.5.0
asgiref==3.5.0
click==8.0.4
colorama==0.4.4
fastapi==0.74.1
h11==0.13.0
httptools==0.3.0
idna==3.3
kafka-python==2.0.2
pydantic==1.9.0
python-dotenv==0.19.2
PyYAML==6.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.1.1
uvicorn==0.17.5
watchgod==0.7
websockets==10.2
clickhouse-driver==0.2.3
pandas==1.3.5
pytest==7.0.1
requests==2.27.1